actions summend up, changed rest appropriately
since its a bad idea to get seperate action files, because it all
becomes a big mess, i summed them up in a single file. there arent that
many actions so its no problem (there arent going to become that much more too)
