+++ app/containers/AddPlayerContainer.js
@@ -1,7 +1,8 @@
 // kept for legacy, as this file isn't used anymore but may be recycled in the future
 
 import { connect } from 'react-redux'
-import { addPlayer } from '../actions/playerActions.js'
+import { setPlayer } from '../actions'
+
 import AddPlayer from '../components/AddPlayer'
 
 const mapStateToProps = (state) => {
@@ -10,8 +11,8 @@ const mapStateToProps = (state) => {
 
 const mapDispatchToProps = (dispatch) => {
   return {
-    onAddPlayer: (player) => {
-      var p = addPlayer(player)
+    onSetPlayer: (player) => {
+      var p = setPlayer(player)
       dispatch(p)
       return p.name
     }
